Add more data to the Secrets and Bindings list, small cleanup (#28)
* Add new info to the secrets and bindings lists

* Cleanup

* Update resources
