--- core-ui/src/components/Predefined/Create/AddonsConfigurations.create.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { AddonsConfigurations } from './AddonsConfigurations/AddonsConfigurations.js';
-
-export const AddonsConfigurationsCreate = ({ resourceType, ...props }) => (
-  <AddonsConfigurations resourceType={resourceType} {...props} />
-);
