+++ core-ui/src/components/Predefined/Create/AddonsConfigurations/AddonsConfigurations.create.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { AddonsConfigurations } from './AddonsConfigurations.js';
+
+export const AddonsConfigurationsCreate = ({ resourceType, ...props }) => (
+  <AddonsConfigurations resourceType={resourceType} {...props} />
+);
+
+export const ClusterAddonsConfigurationsCreate = ({
+  resourceType,
+  ...props
+}) => <AddonsConfigurations resourceType={resourceType} {...props} />;
