--- core-ui/src/components/Predefined/Create/ClusterAddonsConfigurations.create.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import { AddonsConfigurations } from './AddonsConfigurations/AddonsConfigurations.js';
-
-export const ClusterAddonsConfigurationsCreate = ({
-  resourceType,
-  ...props
-}) => <AddonsConfigurations resourceType={resourceType} {...props} />;
