+++ core-ui/src/components/Predefined/Create/index.js
@@ -1,6 +1,4 @@
 export * from './Namespaces.create.js';
-export * from './AddonsConfigurations.create.js';
-export * from './ClusterAddonsConfigurations.create.js';
-export * from './RoleBindings.create.js';
-export * from './ClusterRoleBindings.create.js';
+export * from './AddonsConfigurations/AddonsConfigurations.create.js';
+export * from './RoleBindings/RoleBindings.create.js';
 export * from './OAuthClients/OAuthClients.create.js';
