+++ core-ui/src/components/Predefined/Details/Roles/Roles.details.js
@@ -4,3 +4,7 @@ import { Rules } from './Rules.js';
 export const RolesDetails = DefaultRenderer => ({ ...otherParams }) => {
   return <DefaultRenderer {...otherParams} customComponents={[Rules]} />;
 };
+
+export const ClusterRolesDetails = DefaultRenderer => ({ ...otherParams }) => {
+  return <DefaultRenderer {...otherParams} customComponents={[Rules]} />;
+};
