+++ core-ui/src/components/Predefined/List/RoleBindings.list.js
@@ -1,0 +1,25 @@
+import React from 'react';
+
+export const RoleBindingsList = DefaultRenderer => ({ ...otherParams }) => {
+  const customColumns = [
+    {
+      header: 'Role Name',
+      value: binding => binding.roleRef.name,
+    },
+  ];
+
+  return <DefaultRenderer customColumns={customColumns} {...otherParams} />;
+};
+
+export const ClusterRoleBindingsList = DefaultRenderer => ({
+  ...otherParams
+}) => {
+  const customColumns = [
+    {
+      header: 'Role Name',
+      value: binding => binding.roleRef.name,
+    },
+  ];
+
+  return <DefaultRenderer customColumns={customColumns} {...otherParams} />;
+};
