+++ core-ui/src/components/Predefined/List/Secrets.list.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export const SecretsList = DefaultRenderer => ({ ...otherParams }) => {
+  const customColumns = [
+    {
+      header: 'Type',
+      value: secret => {
+        return secret.type;
+      },
+    },
+  ];
+
+  return <DefaultRenderer customColumns={customColumns} {...otherParams} />;
+};
