+++ core-ui/src/components/Predefined/List/index.js
@@ -4,4 +4,6 @@ export * from './Function/Functions.list.js';
 export * from './Application/Applications.list.js';
 export * from './GitRepository/GitRepositories.list.js';
 export * from './Namespaces.list.js';
+export * from './Secrets.list.js';
+export * from './RoleBindings.list.js';
 export * from './Deployments.list.js';
